feat(client): profile api 외부 주입으로 변경
