+++ apps/client/src/app/profile/create/page.tsx
@@ -2,6 +2,8 @@
 
 import styled from '@emotion/styled';
 import { Button, Spacer, Text } from '@sickgyun/ui';
+import { useQueryClient } from '@tanstack/react-query';
+import { useRouter } from 'next/navigation';
 import type { SubmitHandler } from 'react-hook-form';
 import { useForm } from 'react-hook-form';
 import Footer from '@/components/common/Footer';
@@ -10,17 +12,34 @@ import ProfileForm from '@/components/profile/ProfileForm';
 import { withAuth } from '@/hocs/withAuth';
 import type { CreateProfileRequest } from '@/hooks/api/profile/useCreateProfile';
 import { useCreateProfile } from '@/hooks/api/profile/useCreateProfile';
+import { PROFILE_LIST_QUERY_KEY } from '@/hooks/api/profile/useGetProfileList';
+import { PROFILE_MINE_QUERY_KEY } from '@/hooks/api/profile/useGetProfileMine';
+import { USER_QUERY_KEY } from '@/hooks/api/user/useGetUser';
+import { useLogAnalyticsEvent } from '@/hooks/common/useLogAnalyticsEvent';
 import { useUser } from '@/hooks/common/useUser';
 
 const ProfileCreatePage = () => {
+  const router = useRouter();
   const { user } = useUser();
+  const { logClickEvent } = useLogAnalyticsEvent();
+  const queryClient = useQueryClient();
   const {
     register,
     handleSubmit: handleCreateProfileSubmit,
     setValue,
     watch,
   } = useForm<CreateProfileRequest>();
-  const { mutate: createProfileMutate } = useCreateProfile(watch());
+  const { mutate: createProfileMutate } = useCreateProfile({
+    onSuccess: () => {
+      const profileForm = watch();
+
+      logClickEvent({ name: 'click_create_profile', params: profileForm });
+      queryClient.invalidateQueries({ queryKey: [PROFILE_LIST_QUERY_KEY] });
+      queryClient.invalidateQueries({ queryKey: [PROFILE_MINE_QUERY_KEY] });
+      queryClient.invalidateQueries({ queryKey: [USER_QUERY_KEY] });
+      router.replace('/profile');
+    },
+  });
 
   const onCreateProfile: SubmitHandler<CreateProfileRequest> = (data) => {
     const profile = { isGraduated: user.isGraduated, ...data };
