+++ apps/client/src/app/profile/update/page.tsx
@@ -2,18 +2,29 @@
 
 import styled from '@emotion/styled';
 import { Button, Spacer, Spinner, Text } from '@sickgyun/ui';
+import { useQueryClient } from '@tanstack/react-query';
+import { useRouter } from 'next/navigation';
 import type { SubmitHandler } from 'react-hook-form';
 import { useForm } from 'react-hook-form';
 import Footer from '@/components/common/Footer';
 import Header from '@/components/common/Header';
 import ProfileForm from '@/components/profile/ProfileForm';
 import { withAuth } from '@/hocs/withAuth';
-import { useGetProfileMine } from '@/hooks/api/profile/useGetProfileMine';
+import { PROFILE_QUERY_KEY } from '@/hooks/api/profile/useGetProfile';
+import { PROFILE_LIST_QUERY_KEY } from '@/hooks/api/profile/useGetProfileList';
+import {
+  PROFILE_MINE_QUERY_KEY,
+  useGetProfileMine,
+} from '@/hooks/api/profile/useGetProfileMine';
 import { useUpdateProfile } from '@/hooks/api/profile/useUpdateProfile';
 import type { UpdateProfileRequest } from '@/hooks/api/profile/useUpdateProfile';
+import { useLogAnalyticsEvent } from '@/hooks/common/useLogAnalyticsEvent';
 import { useUser } from '@/hooks/common/useUser';
 
 const ProfileUpdatePage = () => {
+  const router = useRouter();
+  const queryClient = useQueryClient();
+  const { logClickEvent } = useLogAnalyticsEvent();
   const { user, isLoading } = useUser();
   const {
     register,
@@ -21,7 +32,17 @@ const ProfileUpdatePage = () => {
     setValue,
     watch,
   } = useForm<UpdateProfileRequest>();
-  const { mutate: updateProfileMutate } = useUpdateProfile(watch());
+  const { mutate: updateProfileMutate } = useUpdateProfile({
+    onSuccess: () => {
+      const profileForm = watch();
+
+      logClickEvent({ name: 'click_update_profile', params: profileForm });
+      queryClient.invalidateQueries({ queryKey: [PROFILE_LIST_QUERY_KEY] });
+      queryClient.invalidateQueries({ queryKey: [PROFILE_MINE_QUERY_KEY] });
+      queryClient.invalidateQueries({ queryKey: [PROFILE_QUERY_KEY] });
+      router.replace('/profile');
+    },
+  });
   const { profileMine } = useGetProfileMine();
 
   const onUpdateProfile: SubmitHandler<UpdateProfileRequest> = (data) => {
