+++ apps/client/src/components/profile/ProfileDeleteConfirm/index.tsx
@@ -1,5 +1,11 @@
 import { Confirm } from '@sickgyun/ui';
+import { useQueryClient } from '@tanstack/react-query';
+import { useRouter } from 'next/navigation';
 import { useDeleteProfile } from '@/hooks/api/profile/useDeleteProfileMine';
+import { PROFILE_LIST_QUERY_KEY } from '@/hooks/api/profile/useGetProfileList';
+import { PROFILE_MINE_QUERY_KEY } from '@/hooks/api/profile/useGetProfileMine';
+import { USER_QUERY_KEY } from '@/hooks/api/user/useGetUser';
+import { useLogAnalyticsEvent } from '@/hooks/common/useLogAnalyticsEvent';
 
 type ProfileDeleteConfirmProps = {
   onProfileDetailModalClose: VoidFunction;
@@ -10,7 +16,19 @@ const ProfileDeleteConfirm = ({
   onClose,
   onProfileDetailModalClose,
 }: ProfileDeleteConfirmProps) => {
-  const { mutate: deleteProfileMutate } = useDeleteProfile();
+  const router = useRouter();
+  const queryClient = useQueryClient();
+  const { logClickEvent } = useLogAnalyticsEvent();
+
+  const { mutate: deleteProfileMutate } = useDeleteProfile({
+    onSuccess: () => {
+      logClickEvent({ name: 'click_delete_profile_mine' });
+      queryClient.invalidateQueries({ queryKey: [PROFILE_LIST_QUERY_KEY] });
+      queryClient.invalidateQueries({ queryKey: [PROFILE_MINE_QUERY_KEY] });
+      queryClient.invalidateQueries({ queryKey: [USER_QUERY_KEY] });
+      router.replace('/profile');
+    },
+  });
 
   const handleProfileMineDelete = () => {
     deleteProfileMutate();
